Adds routes for signup, login and welcome
