+++ models/user_model.py
@@ -1,0 +1,14 @@
+import sys
+sys.path.append('../')
+
+from app import db, bcrypt
+
+class User(db.Model):
+  __tablename__ = 'users'
+  id = db.Column(db.Integer, primary_key=True)
+  username = db.Column(db.Text, unique=True)
+  password = db.Column(db.Text)
+  
+  def __init__(self, username, password):
+    self.username = username
+    self.password = bcrypt.generate_password_hash(password).decode('UTF-8')
