+++ templates/users/login.html
@@ -1,0 +1,24 @@
+<form method="POST" action="/login">
+    {{ form.csrf_token }}
+    <p>{{ form.username(placeholder="username") }}
+      <span>
+        {% if form.username.errors %}
+          {% for error in form.username.errors %}
+            {{ error }}
+          {% endfor %}
+        {% endif %}
+      </span>
+    </p>
+    <p>
+      {{ form.password(placeholder="password") }}
+      <span>
+        {% if form.password.errors %}
+          {% for error in form.password.errors %}
+            {{ error }}
+          {% endfor %}
+        {% endif %}
+      </span>
+    </p>
+    <button type="submit">Log In!</button>
+</form>
+
