+++ templates/users/welcome.html
@@ -1,0 +1,1 @@
+<h1>You are logged in!</h1>
